RAC-80 rename: dto.res 패키지로 이동
